Python Project/Productivity Web-App/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.103.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
